Drop the deeper-dive path from research onboarding
For v1 we cut the "do a deeper dive" fork: a second research turn that
overwrote the user's already-curated claims and suggestions (and was the
source of the prompt-sync surface Codex flagged). The editable card and the
removal correction stay.

- Reveal suggestions as soon as the result settles instead of gating them
  behind the fork.
- Continue now sends the removal correction (when claims were removed)
  before exiting. exitFocus() clears pending follow-ups, so the correction
  is requested after it; ActiveChatView stays mounted on the same
  conversation and sends it once idle.
- Remove the deeper-dive card + props from the shared view and mock page,
  and the now-dead frozen / deeperDiveResolved / divePending state.
- Rename deeper-dive-prompt.ts -> removal-note-prompt.ts (only buildRemovalNote
  remains) and drop buildDeeperDivePrompt.
