+++ clients/web/src/domains/chat/onboarding-research/research-results-overlay.tsx
@@ -8,11 +8,12 @@
  * flag is set. It reads the conversation transcript from the chat session
  * store and parses the assistant's `{ claims, suggestions }` reply
  * *incrementally* — each claim and suggestion surfaces as its element finishes
- * streaming. A "dig deeper" action sends a follow-up research turn whose result
- * overwrites both. Source favicons resolve from the live web-search results.
+ * streaming. Source favicons resolve from the live web-search results.
  *
- * The underlying chat is intentionally covered; Continue lands the user on the
- * same conversation. Lives in the chat domain because it reads chat state.
+ * The user can remove claims that are wrong; on Continue we fire a one-shot
+ * correction so the assistant disregards them. The underlying chat is
+ * intentionally covered; Continue lands the user on the same conversation.
+ * Lives in the chat domain because it reads chat state.
  */
 
 import { useEffect, useMemo, useState } from "react";
@@ -27,16 +28,14 @@ import { routes } from "@/utils/routes";
 import { ResearchActivityFeed } from "@/domains/chat/onboarding-research/research-activity-feed";
 import { ResearchResultsView } from "@/domains/chat/onboarding-research/research-results-view";
 import {
-  buildDeeperDivePrompt,
   buildRemovalNote,
   type RemovedClaim,
-} from "@/domains/chat/onboarding-research/deeper-dive-prompt";
+} from "@/domains/chat/onboarding-research/removal-note-prompt";
 import {
   extractMessageText,
   latestAssistantMessage,
   parseResearchResultStreaming,
   type RemovalReason,
-  type ResearchFact,
 } from "@/domains/chat/onboarding-research/research-facts";
 
 /** Most claims to show — keeps the card from becoming a wall of text. */
@@ -79,16 +78,8 @@ export function ResearchResultsOverlay() {
     [assistantText],
   );
 
-  // Frozen snapshot — set when the user accepts the result ("This is good for
-  // now"). The follow-up correction we then send gets a non-JSON reply that
-  // would otherwise re-parse to nothing and wipe the card, so once frozen we
-  // render the snapshot instead of the live parse.
-  const [frozen, setFrozen] = useState<{
-    claims: ResearchFact[];
-    suggestions: string[];
-  } | null>(null);
-  const claims = frozen ? frozen.claims : live.claims;
-  const suggestions = frozen ? frozen.suggestions : live.suggestions;
+  const claims = live.claims;
+  const suggestions = live.suggestions;
   const hasFacts = claims.length > 0;
 
   // Removed claims: index → reason (null until/unless one is chosen). Rows stay
@@ -158,34 +149,21 @@ export function ResearchResultsOverlay() {
     return () => clearTimeout(t);
   }, [emptyCandidate]);
 
-  // The deeper-dive prompt is resolved once the user picks either option; the
-  // card then goes away for good and the suggestions are revealed. Sticky.
-  const [deeperDiveResolved, setDeeperDiveResolved] = useState(false);
-
-  // Deeper-dive: bridge the gap between click and the turn actually starting so
-  // the button can't be double-fired.
-  const [divePending, setDivePending] = useState(false);
-  useEffect(() => {
-    if (processing) setDivePending(false);
-  }, [processing]);
-  const handleDeeperDive = () => {
-    if (processing || divePending) return;
+  // Continue out of the focused flow into the same conversation. If the user
+  // removed any claims, fire a one-shot correction so the assistant disregards
+  // them. exitFocus() clears any pending follow-up, so request the correction
+  // *after* exiting — the underlying ActiveChatView stays mounted on this
+  // conversation and sends it once the turn is idle.
+  const handleContinue = () => {
     const removed = collectRemoved();
-    setDivePending(true);
-    setDeeperDiveResolved(true);
-    setRemovals(new Map()); // a fresh result set is coming — drop stale removals
-    requestFollowup(buildDeeperDivePrompt(removed));
-  };
-  const handleGoodForNow = () => {
-    const removed = collectRemoved();
-    // Lock the current result so the correction's reply can't wipe it.
-    setFrozen({ claims: live.claims, suggestions: live.suggestions });
-    setDeeperDiveResolved(true);
+    exitFocus();
     if (removed.length > 0) requestFollowup(buildRemovalNote(removed));
   };
 
   // Clicking a suggestion starts a fresh conversation with that message sent on
-  // the user's behalf, then drops them out of the focused flow into it.
+  // the user's behalf, then drops them out of the focused flow into it. The
+  // removal correction is intentionally skipped here — it belongs to the
+  // conversation being left behind, not the new one.
   const handleSuggestionClick = (suggestion: string) => {
     const draftId = createDraftConversationId();
     useConversationStore.getState().setActiveConversationId(draftId);
@@ -195,10 +173,7 @@ export function ResearchResultsOverlay() {
     exitFocus();
   };
 
-  const settled = started && !processing && !divePending;
-  // Once frozen (accepted) the result is final even while the background
-  // correction turn runs, so show the settled title and allow continue.
-  const locked = frozen != null;
+  const settled = started && !processing;
   const mode: "loading" | "results" | "empty" = hasFacts
     ? "results"
     : showEmpty
@@ -214,21 +189,18 @@ export function ResearchResultsOverlay() {
         removals={removals}
         suggestions={suggestions}
         resultsTitle={
-          settled || locked
+          settled
             ? "Here's what I know about you. You can remove any that aren't true:"
             : "Getting to know you…"
         }
-        showDeeperDiveCard={settled && !deeperDiveResolved}
-        showSuggestions={deeperDiveResolved && suggestions.length > 0}
-        canContinue={!processing || locked}
+        showSuggestions={settled && suggestions.length > 0}
+        canContinue={!processing}
         resolveFavicon={resolveFavicon}
         onRemove={handleRemove}
         onSetReason={handleSetReason}
         onRestore={handleRestore}
-        onDeeperDive={handleDeeperDive}
-        onGoodForNow={handleGoodForNow}
         onSuggestionClick={handleSuggestionClick}
-        onContinue={exitFocus}
+        onContinue={handleContinue}
       />
     </div>
   );
