+++ clients/web/src/domains/chat/onboarding-research/research-results-view.tsx
@@ -29,15 +29,12 @@ export interface ResearchResultsViewProps {
   removals: ReadonlyMap<number, RemovalReason | null>;
   suggestions: string[];
   resultsTitle: string;
-  showDeeperDiveCard: boolean;
   showSuggestions: boolean;
   canContinue: boolean;
   resolveFavicon: (domain: string) => string;
   onRemove: (index: number) => void;
   onSetReason: (index: number, reason: RemovalReason) => void;
   onRestore: (index: number) => void;
-  onDeeperDive: () => void;
-  onGoodForNow: () => void;
   onSuggestionClick: (suggestion: string) => void;
   onContinue: () => void;
 }
@@ -49,15 +46,12 @@ export function ResearchResultsView({
   removals,
   suggestions,
   resultsTitle,
-  showDeeperDiveCard,
   showSuggestions,
   canContinue,
   resolveFavicon,
   onRemove,
   onSetReason,
   onRestore,
-  onDeeperDive,
-  onGoodForNow,
   onSuggestionClick,
   onContinue,
 }: ResearchResultsViewProps) {
@@ -76,21 +70,6 @@ export function ResearchResultsView({
                 resolveFavicon={resolveFavicon}
                 title={resultsTitle}
               />
-              {showDeeperDiveCard ? (
-                <div className="flex flex-col items-start gap-3 rounded-xl border border-[var(--border-base)] bg-[var(--surface-lift)] px-5 py-4">
-                  <p className="text-[15px] text-[var(--content-secondary)]">
-                    I can search more, just didn&apos;t want to be too intrusive.
-                  </p>
-                  <div className="flex flex-wrap gap-2">
-                    <Button variant="primary" size="regular" onClick={onDeeperDive}>
-                      Yes, do a deeper dive
-                    </Button>
-                    <Button variant="outlined" size="regular" onClick={onGoodForNow}>
-                      This is good for now
-                    </Button>
-                  </div>
-                </div>
-              ) : null}
               {showSuggestions ? (
                 <ResearchSuggestions
                   suggestions={suggestions}
